Simplified TTS requests; added List Voices request
